fix missing quoting level in printf format string
the purpose of the inner printf is to create a string of escapes for
the outer printf to act on; as such, the \ should be quoted to be
treated as a literal rather than allowed to bind to the adjacent x.
busybox printf silently ignores the error and treats it as a literal;
bash also treats it as a literal, but produces warnings on stderr.